TOS-23. всё кроме проверки пароля при создании юзера

+++ src/main/java/com/tennecotecnic/onlineshop/controller/UserController.java
@@ -1,40 +1,56 @@
 package com.tennecotecnic.onlineshop.controller;
 
 import com.tennecotecnic.onlineshop.model.Buyer;
-import com.tennecotecnic.onlineshop.repository.UserFileRepository;
-import com.tennecotecnic.onlineshop.repository.UserInMemoryRepository;
 import com.tennecotecnic.onlineshop.repository.UserRepository;
 import com.tennecotecnic.onlineshop.util.PrintUtil;
 import java.io.IOException;
+import com.tennecotecnic.onlineshop.service.LoginService;
 import static com.tennecotecnic.onlineshop.OnlineShop.objectMapper;
 
 public class UserController {
+    private UserRepository userRepository;
+    private LoginService loginService;
 
-    private final UserRepository userRepository = new UserFileRepository();
-    //private final UserRepository userRepository = new UserInMemoryRepository();
+    public UserController(UserRepository userRepository, LoginService loginService){
+        this.userRepository = userRepository;
+        this.loginService = loginService;
+    }
 
     public void processCommand(String stringFromReader) throws IOException {
         String[] commandAndArgument = stringFromReader.split("\\?");
+
         switch (commandAndArgument[0]) {
-            case ("user/create"):
-                Buyer buyer = objectMapper.readValue(commandAndArgument[1], Buyer.class);
-                userRepository.create(buyer);
-                break;
-            case ("user/getAll"):
-                PrintUtil.printUsers(userRepository.findAll());
-                break;
-            case ("user/get"):
-                System.out.println(userRepository.findById(Integer.parseInt(commandAndArgument[1])));
-                break;
-            case ("user/delete"):
+            case ("user/create") -> {
+                if (loginService.checkLogin(commandAndArgument[2])
+                        && loginService.checkAdminRole(commandAndArgument[2])) {
+                    Buyer buyer = objectMapper.readValue(commandAndArgument[1], Buyer.class);
+                    userRepository.create(buyer);
+                }
+            }
+            case ("user/getAll") -> {
+                    if (loginService.checkLogin(commandAndArgument[1])
+                            && loginService.checkAdminRole(commandAndArgument[1])) {
+                    PrintUtil.print(userRepository.findAll());
+                    }
+            }
+            case ("user/get") -> {
+                if (loginService.checkLogin(commandAndArgument[2])
+                        && loginService.checkAdminRole(commandAndArgument[2])) {
+                    System.out.println(userRepository.findById(Integer.parseInt(commandAndArgument[1])));
+                }
+            }
+            case ("user/delete") -> {
+                if (loginService.checkLogin(commandAndArgument[2])) {
                 userRepository.delete(Integer.parseInt(commandAndArgument[1]));
-                break;
-            case ("user/update"):
-                Buyer buyer1 = objectMapper.readValue(commandAndArgument[1], Buyer.class);
-                userRepository.update(buyer1);
-                break;
-            default:
-                System.out.println("Invalid command for UsersRepository");
+                }
+            }
+            case ("user/update") -> {
+                if (loginService.checkLogin(commandAndArgument[2])) {
+                    Buyer buyer1 = objectMapper.readValue(commandAndArgument[1], Buyer.class);
+                    userRepository.update(buyer1);
+                }
+            }
+            default -> System.out.println("Invalid command for UsersRepository");
         }
     }
 }
